+++ structs/gateway/session/udp_session.go
@@ -85,7 +85,6 @@ func NewUdpSession() UdpSession {
 	u.SetEventDecoders(&payload.DiscoveryPacket{}, &payload.VoicePacket{}, &payload.SenderReportPacket{})
 	u.SetErrorHandlers(map[error]func(){
 		net.ErrClosed: func() {
-			u.Exit(false)
 		},
 		io.EOF: func() {
 			u.Exit(false)